refactor: rename files from pascalcase to kebab-case
