+++ src/app/search/[startDate]/[endDate]/[apartmentId]/[bookingSessionId]/book/page.tsx
@@ -1,4 +1,4 @@
-import WireframeSection from "@/components/wireframe/WireframeSection";
+import WireframeSection from "@/components/wireframe/wireframe-section";
 
 const BookingFunnelPersonalDetails = () => {
 	return (
