+++ src/app/search/[startDate]/[endDate]/[apartmentId]/[bookingSessionId]/pay/page.tsx
@@ -1,4 +1,4 @@
-import WireframeSection from "@/components/wireframe/WireframeSection";
+import WireframeSection from "@/components/wireframe/wireframe-section";
 
 const BookingFunnelPaymentDetails = () => {
 	return (
