+++ src/app/search/[startDate]/[endDate]/[apartmentId]/page.tsx
@@ -1,4 +1,4 @@
-import WireframeSection from "@/components/wireframe/WireframeSection";
+import WireframeSection from "@/components/wireframe/wireframe-section";
 
 const ApartmentDetailPage = () => {
 	return (
